counters: cast func counter values to the signed stats domain
StatsRecord value is int64_t while counter callbacks return uint64_t.
Cast explicitly: values above INT64_MAX (e.g. an underflowed memuse
counter) tripped -fsanitize=implicit-conversion in the stats thread.

+++ src/counters.c
@@ -792,7 +792,7 @@ static int StatsOutput(ThreadVars *tv)
             switch (pc->type) {
                 case STATS_TYPE_FUNC:
                     if (pc->Func != NULL)
-                        thread_table[pc->gid].value = pc->Func();
+                        thread_table[pc->gid].value = (int64_t)pc->Func();
                     break;
                 case STATS_TYPE_AVERAGE:
                     DEBUG_VALIDATE_BUG_ON(i + 1 >= max_id); // help scan-build
